Fix carry_regime hardcoding, reset leak, and float type annotations
carry_regime_cuts now loaded lazily from affinity_scores.json when not
explicitly passed — no hardcoded fallback. target_rr cleared on reset().
Scorer tuple type corrected to float to match fractional JSON scores.

+++ knowledge_engine/structure_scorer.py
@@ -36,8 +36,8 @@ def score_structures(
 
     buckets = _compute_buckets(market_state, thresholds)
 
-    primary: list[tuple[int, str]] = []   # (total_score, struct_id)
-    overlays: list[tuple[int, str]] = []
+    primary: list[tuple[float, str]] = []
+    overlays: list[tuple[float, str]] = []
 
     for struct_id, score_cfg in struct_scores.items():
         if struct_id not in profiles:
